+++ .gitignore
@@ -1,7 +1,3 @@
 *.pyc
-wav
-mp3
-*.wav
-*.mp3
 .DS_Store
 *.cnf
