+++ dejavu/base_classes/base_recognizer.py
@@ -1,0 +1,19 @@
+import abc
+
+from dejavu.config.settings import DEFAULT_FS
+
+
+class BaseRecognizer(object, metaclass=abc.ABCMeta):
+    def __init__(self, dejavu):
+        self.dejavu = dejavu
+        self.Fs = DEFAULT_FS
+
+    def _recognize(self, *data):
+        matches = []
+        for d in data:
+            matches.extend(self.dejavu.find_matches(d, Fs=self.Fs))
+        return self.dejavu.align_matches(matches)
+
+    @abc.abstractmethod
+    def recognize(self):
+        pass  # base class does nothing
