--- dejavu/database_handler/postgres_queries.py
@@ -1,104 +1,0 @@
-from dejavu.config.config import (FIELD_FILE_SHA1, FIELD_FINGERPRINTED,
-                                  FIELD_HASH, FIELD_OFFSET, FIELD_SONG_ID,
-                                  FIELD_SONGNAME, FINGERPRINTS_TABLENAME,
-                                  SONGS_TABLENAME)
-
-# creates
-CREATE_SONGS_TABLE = f"""
-    CREATE TABLE IF NOT EXISTS "{SONGS_TABLENAME}" (
-        "{FIELD_SONG_ID}" SERIAL
-    ,   "{FIELD_SONGNAME}" VARCHAR(250) NOT NULL
-    ,   "{FIELD_FINGERPRINTED}" SMALLINT DEFAULT 0
-    ,   "{FIELD_FILE_SHA1}" BYTEA
-    ,   "date_created" TIMESTAMP NOT NULL DEFAULT now()
-    ,   "date_modified" TIMESTAMP NOT NULL DEFAULT now()
-    ,   CONSTRAINT "pk_{SONGS_TABLENAME}_{FIELD_SONG_ID}" PRIMARY KEY ("{FIELD_SONG_ID}")
-    ,   CONSTRAINT "uq_{SONGS_TABLENAME}_{FIELD_SONG_ID}" UNIQUE ("{FIELD_SONG_ID}")
-    );
-"""
-
-CREATE_FINGERPRINTS_TABLE = f"""
-    CREATE TABLE IF NOT EXISTS "{FINGERPRINTS_TABLENAME}" (
-        "{FIELD_HASH}" BYTEA NOT NULL
-    ,   "{FIELD_SONG_ID}" INT NOT NULL
-    ,   "{FIELD_OFFSET}" INT NOT NULL
-    ,   "date_created" TIMESTAMP NOT NULL DEFAULT now()
-    ,   "date_modified" TIMESTAMP NOT NULL DEFAULT now()
-    ,   CONSTRAINT "uq_{FINGERPRINTS_TABLENAME}" UNIQUE  ("{FIELD_SONG_ID}", "{FIELD_OFFSET}", "{FIELD_HASH}")
-    ,   CONSTRAINT "fk_{FINGERPRINTS_TABLENAME}_{FIELD_SONG_ID}" FOREIGN KEY ("{FIELD_SONG_ID}") 
-            REFERENCES "{SONGS_TABLENAME}"("{FIELD_SONG_ID}") ON DELETE CASCADE
-    );
-
-    CREATE INDEX IF NOT EXISTS "ix_{FINGERPRINTS_TABLENAME}_{FIELD_HASH}" ON "{FINGERPRINTS_TABLENAME}" USING hash ("{FIELD_HASH}");
-"""
-
-CREATE_FINGERPRINTS_TABLE_INDEX = f"""
-    CREATE INDEX "ix_{FINGERPRINTS_TABLENAME}_{FIELD_HASH}" ON "{FINGERPRINTS_TABLENAME}" USING hash ("{FIELD_HASH}");
-"""
-
-# inserts (ignores duplicates)
-INSERT_FINGERPRINT = f"""
-    INSERT INTO "{FINGERPRINTS_TABLENAME}" (
-            "{FIELD_SONG_ID}"
-        ,   "{FIELD_HASH}"
-        ,   "{FIELD_OFFSET}") 
-    VALUES (%s, decode(%s, 'hex'), %s) ON CONFLICT DO NOTHING;
-"""
-
-INSERT_SONG = f"""
-    INSERT INTO "{SONGS_TABLENAME}" ("{FIELD_SONGNAME}", "{FIELD_FILE_SHA1}") 
-    VALUES (%s, decode(%s, 'hex'))
-    RETURNING "{FIELD_SONG_ID}";
-"""
-
-# selects
-SELECT = f"""
-    SELECT "{FIELD_SONG_ID}", "{FIELD_OFFSET}" 
-    FROM "{FINGERPRINTS_TABLENAME}"
-    WHERE "{FIELD_HASH}" = decode(%s, 'hex');
-"""
-
-SELECT_MULTIPLE = f"""
-    SELECT upper(encode("{FIELD_HASH}", 'hex')), "{FIELD_SONG_ID}", "{FIELD_OFFSET}" 
-    FROM "{FINGERPRINTS_TABLENAME}"
-    WHERE "{FIELD_HASH}" IN (%s);
-"""
-
-SELECT_ALL = f'SELECT "{FIELD_SONG_ID}", "{FIELD_OFFSET}" FROM "{FINGERPRINTS_TABLENAME}";'
-
-SELECT_SONG = f"""
-    SELECT "{FIELD_SONGNAME}", upper(encode("{FIELD_FILE_SHA1}", 'hex')) AS "{FIELD_FILE_SHA1}" 
-    FROM "{SONGS_TABLENAME}"
-    WHERE "{FIELD_SONG_ID}" = %s;
-"""
-
-SELECT_NUM_FINGERPRINTS = f'SELECT COUNT(*) AS n FROM "{FINGERPRINTS_TABLENAME}";'
-
-SELECT_UNIQUE_SONG_IDS = f"""
-    SELECT COUNT("{FIELD_SONG_ID}") AS n 
-    FROM "{SONGS_TABLENAME}"
-    WHERE "{FIELD_FINGERPRINTED}" = 1;
-"""
-
-SELECT_SONGS = f"""
-    SELECT 
-        "{FIELD_SONG_ID}"
-    ,   "{FIELD_SONGNAME}"
-    ,   upper(encode("{FIELD_FILE_SHA1}", 'hex')) AS "{FIELD_FILE_SHA1}" 
-    FROM "{SONGS_TABLENAME}" 
-    WHERE "{FIELD_FINGERPRINTED}" = 1;
-"""
-
-# drops
-DROP_FINGERPRINTS = f'DROP TABLE IF EXISTS "{FINGERPRINTS_TABLENAME}";'
-DROP_SONGS = f'DROP TABLE IF EXISTS "{SONGS_TABLENAME}";'
-
-# update
-UPDATE_SONG_FINGERPRINTED = f"""
-    UPDATE "{SONGS_TABLENAME}" SET "{FIELD_FINGERPRINTED}" = 1, "date_modified" = now() WHERE "{FIELD_SONG_ID}" = %s;
-"""
-
-# delete
-DELETE_UNFINGERPRINTED = f"""
-    DELETE FROM "{SONGS_TABLENAME}" WHERE "{FIELD_FINGERPRINTED}" = 0;
-"""
