+++ dejavu/logic/fingerprint.py
@@ -1,0 +1,108 @@
+import hashlib
+from operator import itemgetter
+
+import matplotlib.mlab as mlab
+import matplotlib.pyplot as plt
+import numpy as np
+from scipy.ndimage.filters import maximum_filter
+from scipy.ndimage.morphology import (binary_erosion,
+                                      generate_binary_structure,
+                                      iterate_structure)
+
+from dejavu.config.settings import (DEFAULT_AMP_MIN, DEFAULT_FAN_VALUE,
+                                    DEFAULT_FS, DEFAULT_OVERLAP_RATIO,
+                                    DEFAULT_WINDOW_SIZE, FINGERPRINT_REDUCTION,
+                                    MAX_HASH_TIME_DELTA, MIN_HASH_TIME_DELTA,
+                                    PEAK_NEIGHBORHOOD_SIZE, PEAK_SORT)
+
+IDX_FREQ_I = 0
+IDX_TIME_J = 1
+
+
+def fingerprint(channel_samples,
+                Fs=DEFAULT_FS,
+                wsize=DEFAULT_WINDOW_SIZE,
+                wratio=DEFAULT_OVERLAP_RATIO,
+                fan_value=DEFAULT_FAN_VALUE,
+                amp_min=DEFAULT_AMP_MIN):
+    """
+    FFT the channel, log transform output, find local maxima, then return locally sensitive hashes.
+    """
+    # FFT the signal and extract frequency components
+    arr2D = mlab.specgram(
+        channel_samples,
+        NFFT=wsize,
+        Fs=Fs,
+        window=mlab.window_hanning,
+        noverlap=int(wsize * wratio))[0]
+
+    # Apply log transform since specgram() returns linear array. 0s are excluded to avoid np warning.
+    arr2D = 10 * np.log10(arr2D, out=np.zeros_like(arr2D), where=(arr2D != 0))
+
+    local_maxima = get_2D_peaks(arr2D, plot=False, amp_min=amp_min)
+
+    # return hashes
+    return generate_hashes(local_maxima, fan_value=fan_value)
+
+
+def get_2D_peaks(arr2D, plot=False, amp_min=DEFAULT_AMP_MIN):
+    #  http://docs.scipy.org/doc/scipy/reference/generated/scipy.ndimage.iterate_structure.html#scipy.ndimage.iterate_structure
+    struct = generate_binary_structure(2, 1)
+    neighborhood = iterate_structure(struct, PEAK_NEIGHBORHOOD_SIZE)
+
+    # find local maxima using our filter shape
+    local_max = maximum_filter(arr2D, footprint=neighborhood) == arr2D
+    background = (arr2D == 0)
+    eroded_background = binary_erosion(background, structure=neighborhood, border_value=1)
+
+    # Boolean mask of arr2D with True at peaks (Fixed deprecated boolean operator by changing '-' to '^')
+    detected_peaks = local_max ^ eroded_background
+
+    # extract peaks
+    amps = arr2D[detected_peaks]
+    freqs, times = np.where(detected_peaks)
+
+    # filter peaks
+    amps = amps.flatten()
+    # get indices for frequency and time
+    filter_idxs = np.where(amps > amp_min)
+
+    freqs_filter = freqs[filter_idxs]
+    times_filter = times[filter_idxs]
+
+    if plot:
+        # scatter of the peaks
+        fig, ax = plt.subplots()
+        ax.imshow(arr2D)
+        ax.scatter(times_filter, freqs_filter)
+        ax.set_xlabel('Time')
+        ax.set_ylabel('Frequency')
+        ax.set_title("Spectrogram")
+        plt.gca().invert_yaxis()
+        plt.show()
+
+    return list(zip(freqs_filter, times_filter))
+
+
+def generate_hashes(peaks, fan_value=DEFAULT_FAN_VALUE):
+    """
+    Hash list structure:
+       sha1_hash[0:20]    time_offset
+    [(e05b341a9b77a51fd26, 32), ... ]
+    """
+    if PEAK_SORT:
+        peaks.sort(key=itemgetter(1))
+
+    for i in range(len(peaks)):
+        for j in range(1, fan_value):
+            if (i + j) < len(peaks):
+
+                freq1 = peaks[i][IDX_FREQ_I]
+                freq2 = peaks[i + j][IDX_FREQ_I]
+                t1 = peaks[i][IDX_TIME_J]
+                t2 = peaks[i + j][IDX_TIME_J]
+                t_delta = t2 - t1
+
+                if MIN_HASH_TIME_DELTA <= t_delta <= MAX_HASH_TIME_DELTA:
+                    h = hashlib.sha1(f"{str(freq1)}|{str(freq2)}|{str(t_delta)}".encode('utf-8'))
+                    yield (h.hexdigest()[0:FINGERPRINT_REDUCTION], t1)
