+++ setup.cfg
@@ -1,0 +1,3 @@
+[flake8]
+max-line-length = 120
+
